# Store payload as String for faster search/analyzing