Logs searching and timeline analyzing happen more often than file parsing. In this case it makes sense to store payload as text - this slows down initial file parsing but on another hand drastically improves working with parsed  messages.
Theoretically, memory consumption will be lower with string payload representation.

1GB DLT file test (7.156.519 messages)
|  | Parsing | Searching | Analyzing | Total time |
| - | - | - | - | - |
| Raw payload | 5s | 71s | 28s | 104s |
| String payload | 43s | 19s | 24s | 86s |
| Raw payload + progress update frequency limit | 4s | 68s | 19s | 91s |
| String payload + progress update frequency limit | 34s | 17s | 17s | 68s |